backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
joblib==1.3.2
pydantic==2.6.0
python-multipart
httpx==0.24.1
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
mlflow
xgboost
statsmodels
pandas
numpy
typer
rich
httpx
joblib
python-multipart